requirements.txt
pyaudio
opencv-python
pillow
ffmpeg
transformers
SpeechRecognition
deep-translator
gTTS
playsound